test all attributes, Update Astrophotographer index page, route and action

+++ db/migrate/20200926212622_create_constellations.rb
@@ -3,7 +3,7 @@ class CreateConstellations < ActiveRecord::Migration[6.0]
     create_table :constellations do |t|
       t.string :name
       t.string :description
-      t.decimal :light_years_away_from_earth
+      t.float :light_years_away_from_earth
       t.string :image_url
 
       t.timestamps
